// app/shared/models/DocCategory.model.ts
export class DocCategoryModel {
     doccategoryid: number = 0;
     doccategory: string = null;
     doccategoryorder: number = 0;
     is_active: number = 0;
     createdby: number =0;
     doccategorycode:string=null;
     updateby : number =0;
}
export class DocCategoryViewModel 
{
     nDocCategoryId: number = 0;
     tDocCategory: string="";
     nDocCategoryOrder: number = 0;
     dCreatedDt : string = ""
     bIsActive: boolean = false;
}